test(gfx950): separate MXFP quantization coverage
